[#56] showing scale in rendered graph
